Added IRC Announce as per issue #4.
If configured, an Announce via IRC option is presented on the new mosh view.
If one of these options are checked, a GET is done to the specified URL with the
params "channel" and "message", when the channel is that which was checked, and
the message is the one specified in the config with values interpolated.

Also changed the ID schema to a SHA1 Hex, similar to that of GH, rather than a
UUID which is longer than a Donkey's cock.