Added "Raw Mosh" button
As per issue #17, requested by @Abablabab there is now a button
on the mosh view that opens the mosh data in a raw format for saving/downloading
etc.

Also re-worded the README, emptied the dancer-scafold index.tt which
isn't used and fixed a floating issue with the mosh id and mosh time in
wider browsers.